requirements.txt
Faker==30.3.0
iniconfig==2.0.0
numpy==2.1.2
packaging==24.1
pluggy==1.5.0
py4j==0.10.9.7
pyspark==3.5.3
pytest==8.3.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.1.1
six==1.16.0
typing_extensions==4.12.2
